JAVA program to demonstrate Strassens algorithm
Algorithm adapted from CLRS book , this .java program will run on
terminal , and generate a info.log in the directory it is run from.
To generate the log file, call the program with parameter "true" .

The log file contains the trace matrices e.g. S, P and C matrices to
verify the same manually.